# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"ProPhishy backend","version":"0.1.0"},"paths":{"/metrics":{"get":{"summary":"Get Metrics","operationId":"get_metrics_metrics_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/flagged":{"get":{"summary":"Get Flagged","operationId":"get_flagged_flagged_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/refresh":{"post":{"summary":"Refresh Emails","description":"Fetch latest messages from Gmail and classify them. Returns the new flagged list.","operationId":"refresh_emails_refresh_post","parameters":[{"name":"max_results","in":"query","required":false,"schema":{"type":"integer","default":25,"title":"Max Results"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/action/mark-safe":{"post":{"summary":"Mark Safe","description":"Remove an email from flagged (user marks it safe).\npayload: {\"id\": \"<message id>\"}","operationId":"mark_safe_action_mark_safe_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Payload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/action/delete":{"post":{"summary":"Action Delete","description":"Delete (trash) message in Gmail and remove from flagged list.\npayload: {\"id\": \"<message id>\"}","operationId":"action_delete_action_delete_post","requestBody":{"content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Payload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}